requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.0.7
django-crispy-forms==2.3
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==3.0.2
openpyxl==3.1.5
pillow==10.4.0
pytz==2024.1
qrcode==8.2
reportlab==4.2.5
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
